LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_suppress_compat_events_on_click-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Event: Suppress compatibility mouse events on click

When a pointerdown is canceled, a click/tap shouldn't fire any compatibility mouse events except click event.

Click or tap on Target0.
Click or tap on Target1.
Click Done.
Target0
Target1
Done
The following pointer types were detected: (none).

The following events were logged: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN Suppress compat mouse events on click
